+++ source/renderer/app/components/wallet/receive/WalletReceiveRandom.js
@@ -200,6 +200,9 @@ export default class WalletReceiveRandom extends Component<Props, State> {
     ]);
 
     const passwordField = form.$('spendingPassword');
+
+    const canSubmit = !isSubmitting && passwordField.value;
+
     const generateAddressForm = (
       <div className={generateAddressWrapperClasses}>
         {walletHasPassword && (
@@ -217,6 +220,7 @@ export default class WalletReceiveRandom extends Component<Props, State> {
 
         <Button
           className={generateAddressButtonClasses}
+          disabled={!canSubmit}
           label={intl.formatMessage(messages.generateNewAddressButtonLabel)}
           skin={ButtonSkin}
           onClick={this.submit}